backend/requirements.txt
fastapi
uvicorn
python-multipart
python-dotenv
openai
PyPDF2
python-docx
requirememnts